Add movies with just the IMDB id.
I know, this commit is too massive, but each little change led to another and
it was too late to notice I should have made more commits out of this. Some of
the changes are refactoring, probably causing new exceptions in unexpected
places. Tried to test somewhat.

- Can add from command line without movie name. Just need an IMDB id (or url)
- imdbutils and id clean functions throw new exception BadIMDBIdException for
  handling the situations we need correct IMDB id data or nothing.
- Add new IMDB query function for fetching movie names by id.
- Rename some imdbutils functions
- completely rework moary.do_add

Not the best of commits by any standards.