Add first tests.
Python's own unittest and for mocks, stubs and dummies there is the third-party
ludibrio, found here: https://github.com/nsigustavo/ludibrio

These couple of tests here are just for to get the engine warm. The idea is to
have some whiteboxish tests to prevent regression. These particular tests in
this commit are written against uncommited code and can be itself very wrong.
In time we'll see how this helps us.